refactor(rum-legacy): remove duplicated object and error helpers
The merge and empty-check loops existed twice, byte for byte, in event
assembly and in the public api, because Object.assign and the spread
operator both need ES2015 and lib: ES5 rejects them. They move to
tools/objectUtils.ts.

The block reading a message, name and stack off an Error instance also
existed twice in error collection, once for uncaught errors and once
for manually added ones.

No behaviour change.
